+++ src/functions_make_units.py
@@ -7,16 +7,33 @@ from classes_buildings import *
 def make_naval_factories(map, dict_with_game_state, dict_with_units):
 # make naval factories
     id = dict_with_game_state["lowest_free_id"]
-    for place in map.places_for_naval_factories:        
+    for place in map.places_for_naval_factories:    
+        # make factory      
         dict_with_units[id] = Navy_factory(id, place, 0, 0, 0)
+        # set default target for newly produced units (oriented to the center of the map)
+        place_but_in_ids = map.world2id(place)
+        if place_but_in_ids[0] > map.map_width/2: angle = math.pi
+        else: angle = 0
+        # if the center of the map is not water, turn the angle
+        if map.BOARD[map.map_height//2][map.map_width//2].type not in ["water", "shallow"]:
+            angle += math.pi
+        dict_with_units[id].target_for_units = [move_point(place, 200, angle)]
+        # increase id
         id += 1
     dict_with_game_state["lowest_free_id"] = id
 
 def make_land_factories(map, dict_with_game_state, dict_with_units):
 # make land factories
     id = dict_with_game_state["lowest_free_id"]
-    for place in map.places_for_land_factories:        
+    for place in map.places_for_land_factories:      
+        # make factory  
         dict_with_units[id] = Land_factory(id, place, 0, 0, 0)
+        # set default target for newly produced units (oriented to the center of the map)
+        place_but_in_ids = map.world2id(place)
+        if place_but_in_ids[0] > map.map_width/2: angle = math.pi
+        else: angle = 0
+        dict_with_units[id].target_for_units = [move_point(place, 200, angle)]
+        # increase id
         id += 1
     dict_with_game_state["lowest_free_id"] = id